requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.4.0
async-timeout==3.0.1
asyncpraw==7.2.0
asyncprawcore==2.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
fastapi==0.65.1
fastapi-route-logger-middleware==0.1.3
feedparser==6.0.2
h11==0.12.0
idna==2.10
lxml==4.6.3
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.3
openpyxl==3.0.7
pandas==1.3.5
parse==1.19.0
pydantic==1.8.2
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytrends==4.7.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-html==0.10.0
scipy==1.7.3
searchtweets-v2==1.0.7
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
tqdm==4.60.0
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.4
uvicorn==0.14.0
w3lib==1.22.0
websockets==8.1
yarl==1.6.3
